ref: rename Header template to PublicHeader
