+++ frontend/src/app/AppPublicHeader.tsx
@@ -3,7 +3,7 @@ import { Button } from '@chakra-ui/react'
 
 import { BxsHelpCircle } from '~assets/icons/BxsHelpCircle'
 import { LOGIN_ROUTE } from '~constants/routes'
-import { PublicHeader } from '~templates/Header'
+import { PublicHeader } from '~templates/PublicHeader'
 
 const PUBLIC_HEADER_LINKS = [
   {
