+++ frontend/src/templates/PublicHeader/PublicHeader.stories.tsx
